requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0

# Google ADK
google-adk==1.0.0

# LiteLlm for multi-provider LLM support
litellm==1.66.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Authentication
PyJWT==2.8.1
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0